requirements.txt
cffi==1.17.1
cryptography==44.0.0
pillow==11.0.0
pycparser==2.22
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
